github migration: Replace git:// with https://
The git:// interface is obsolete in github and will be removed:
https://github.blog/2021-09-01-improving-git-protocol-security-github/#no-more-unauthenticated-git

Some packages do not support this update. Replacing git:// with https://
triggers the requirement to add sha256 sums, however there is a bug
in the yocto recipetool which causes the fetch task to fail with error:
"Checksum mismatch!
File: '.../autoivnsw_sja1110_linux' has sha256 checksum 175... when b7e...
was expected".
As per advice:
https://www.yoctoproject.org/pipermail/yocto/2015-July/025851.html
will let those packages use git://, following to change that when upgrading
to yocto 4.0.

Issue: ALB-9358
Issue: ALB-9563
Signed-off-by: Ionut Vicovan <Ionut.Vicovan@nxp.com>