nativeaptinstall: Added user preinstall function
Before we start installing a multitude of packages we may have to do
some image specific sanity checks, i.e., we need a preinstall function
that can be called before the major install work commences. This
patch does that split.

Rationale: We may have to do some user/group cleanup similar to what
the Yocto rootfs postinstalls do, but that is image specific so it
shouldn't be done in the class.

Issue: ALB-9499
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>