nativeaptinstall: Better printing of installed packages
For validation of the Ubuntu recipes the class prints the installed
package list in the log. This list needs to be matched and moved into
the recipes which was complicated in the prior form. Now the
architecture suffixes are stripped as they are not helpful and a few
lines surrounding the list are added to make finding it easier in the
log.

Issue: ALB-9499
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>