fsl-image-ubuntu-base: Now properly checks the systemd users/groups
The Yocto code to turn the systemd database into users/groups can't
be used because it doesn't really run with the Ubuntu chroot tools.
It also didn't parse things properly and did not build command lines
properly. So here we have our own code now.

Turns out that it really only works on already existing users/groups
right now and therefore does effectively nothing in the current
Ubuntu setups we support, but we may need it for different rootfs
setups, so we just leave it "as is" for now.

Issue: ALB-9499
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>