imagebuilder: Upgrade to version 1.0
Changed the imagebuilder version to be the official 1.0 release
and to be fetched from the Xen Project imagebuilder repo instead
of the old ViryaOS project.

Also added "dtc-native" to RDEPENDS-native since the newer version
needs it.

Removed the old downstream patches from the imagebuilder recipe
since it will be upgraded to a newer version which already
contains the upstreamed patches.

Added the patch which enables dynamically computing the addresses
and sizes of loaded binaries via u-boot variables instead of
computing them inside imagebuilder and hardcoding them into the
generated script.

Issue: ALB-9661
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>