xen-examples: Update parameters for imagebuilder v1.0
Updated the parameters for running the imagebuilder script
to correspond with the ones expected by imagebuilder v1.0.

As a result of the imagebuilder upgrade, the APPEND_EXTRA_CMDS
option is no longer passed as a parameter to the script, but instead
it is read from the config files passed to the script, via the
"APPEND_EXTRA_CMDS=<path_to_file_with_cmds>" variable.

Added two commands to the extra commands file, to be run at the
end of the script, which store the load addreses for xen and
the device-tree inside "xen_addr" and "fdt_addr", to ensure
compability with the way our u-boot environment is set.

The dynamic addresses and sizes option inside imagebuilder is
no longer used by default, but is enabled via the "-s" parameter
instead, which is now passed to the script.

Also renamed the extra_cmd_s32 file to extra_cmd_s32gen1.

Issue: ALB-9667
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>