xen-examples: dom0less: Set "xen,enhanced" property for PV drivers
Set the "DOMU_ENHANCED" property to enable the support in Xen
for PV drivers for Dom0less VMs. This makes Imagebuilder add the
"xen,enhanced" property to the /chosen DT node for the DomU, but
the setup for any PV drivers has to be done at runtime.

The Dom0less VMs in passthrough setups will have it disabled,
for the moment.

Issue: ALB-9667
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>