xen-passthrough-dts: Use nativepython3 when calling partial-dtb-gen
When the "xen-passthrough-dts" recipe uses the partial-dtb-gen python
tool, it requires the python-fdt package installed. To use the correct
python for native recipes it is needed to use the "nativepython3"
package instead of calling "python3". Otherwise, the python3 binary
called would not find the "python-fdt" package correctly installed for
the native version, and a fallback (try/catch) mechanism in the
"partial-dtb-gen" tool would manually install it on the host, which
we want to avoid.

Thus, resorted to using "nativepython3" for calling
"partial-dtb-gen.py".

Issue: ALB-9729
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>