vim: Strip 'vim' package
As part of 'fsl-image-auto' BSP image support, the
'vim' package is being used. By default, it enabled several
graphical dependencies which are not useful as part of S32CC
platforms.
As part of this commit, we strip those dependencies in order
to reduce the build time of 'fsl-image-auto' BSP image
in Yocto 4.0 ('kirkstone') version.

Issue: ALB-9814
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>