recipes-bsp: hse: add TrustZone Key Provisioning Demo support
The TrustZone Key Provisioning Demo would generate the Key Encryption
Key (KEK) on the host. Aditionally, an symmetric key is also generated
on the host and encrypted & authenticated using the KEK. These keys are
transferred in the target's rootfs for later provisioning in HSE NVM
Key Catalog on the target device.
Besides key generation & encryption, two executables are tranferred in
the target's rootfs to provide the provisioning feature in TrustZone.
Since the TrustZone Key Provisioning feature is done through OP-TEE,
this feature is enabled only when 'optee' and 'hse' distro features
are added.

Issue: ALB-9778
Signed-off-by: Bogdan Roman <bogdan-gabriel.roman@nxp.com>