meta-alb: Use immediate expansion for FILESEXTRAPATHS variable
Replaced the assignment operator with the immediate expansion assignment
one for xen-examples and xen-passthrough-dts recipes, since that is
the standard.

Because the main logic for xen-examples is in a ".inc" file which is
included by other recipes, if another user wanted to define a
xen-examples recipe in another layer (e.g.: meta-gvip) and use it, the
FILESEXTRAPATHS variable would expand to the path within that layer,
which does not contain the extra_cmds_s32cc file, for example, and
it would lead to a build error.

Issue: ALB-9862
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>