ls2080: t4: Removed obsolete references to T4 and ls2080abluebox
We had T4 BlueBox leftovers. We removed those.
We also decided that supporting the ls2080abluebox, aka BlueBox 1, right
now is a problem because the current TF-A does not have easy A57 support
for that board specifically. So we drop the ls2080abluebox support.
Re-adding it is not too hard, but right now it is better not to leave
an illusion.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>