image_types_fsl_itb: bbclass: Option to use non-standard kernels
Usually we want to use virtual/kernel when creating an ITB. However,
there are situations when virtual/kernel is not suitable and a custom
kernel should be used. One such case is when a size reduced kernel for
a flash image is needed while not touching the main virtual/kernel.

This patch introduces some variables that allow selecting a different
kernel as needed.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>