linux-flash: Added support for a size reduced version of virtual/kernel
While we want virtual/kernel for main rootfs and operational setups,
recent kernel sizes pose a problem for flash recovery images with
minimum rootfs.

This new set of recipes builds a size reduced version of virtual/kernel
as linux-flash, including associated modules.

It is not meant to add a single feature, only to remove features to
reduce kernel size, so that a recent kernel and a rootfs can fit into
64MB of flash together with other boot and firmware files.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>