bbclass: Add varflagsremove.bbclass
It implements a remove option for variable flags that Yocto syntax
does not currently support.
The support is generic and can be used by anyone who would need to
do a

var[flag]:remove = ""

type of operation, even though the syntax is of course slightly
different as we cannot change bitbake operation.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>