u-boot-environment: Now support includes that are global
By default the UBOOT_ENV_NAME is searched within the machine directory.
To permit an include that is global to all environments for reuse, we
introduce UBOOT_ENV_INC_NAME, which then contains files that the
environments in the machine dirs may include.

The basic idea here is that common parts of an environment can be
factored out, eliminating a lot of copy/paste.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>