rcw: Splitting up A050426 PBI to keep PBI below 4KB.
The full A050426 workaround in the PBI combined with other workarounds
can cause the PBI to exceed the 4KB supported by the boot ROMs.
So we split it up in subfiles and permit BlueBox3 to only use those
elements that are needed.

That way we can combine A050426 with the necessary SerDes 0x1f config
and the other required elements.

Issue: ALB-10032
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>