hse: HSE Firmware recipe should not introduce bitbake parse issues
The way the recipe "require" statement was written it broke if the
recipe was parsed in a context with an unknown SoC. This is normal if
the layer is used in a more generic context.
To ensure the recipe can be parsed at all times, we require the
included file only if it is defined/it exists.
This way the recipe is no longer a problem and retains
its normal functionality.

Issue: ALB-10439
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>