linux-s32: clean-up the DTB before each kernel compile
After linux's compilation step is done, depending on the configuration
from DISTRO_FEATURES the build system will modify the DTBs
using scmi_dtb_node_change.py script.
Since the script does not modify the input (the .dts file) but the
output (the .dtb file), sequential builds might result in a unbootable board.
This happens because the kernel does not recompile a DTB if it already
exists.

When building the kernel, before compiling it, delete the current used
DTB in the output directory.

Issue: ALB-10459
Signed-off-by: Alexandru-Catalin Ionita <alexandru-catalin.ionita@nxp.com>