github: migration: Eliminate 'fsl-github-migrate' bbclass
At this point, we do not use any codeaurora URL.
Therefore, we can completely remove the
'fsl-github-migrate' bbclass and its related usage from
include files. Removing this hack is a major cleanup.
It also eliminates the use of warnings for dangling bbappends,
as this was part of the include file to use the class.

Issue: ALB-10510
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>