Fix ld.so segmentation fault on native sdk
After installing a native sdk generated with
`bitbake fsl-image-base -c populate_sdk`, any binary from
the sdk was seg faulting.

The reason is that GCC11 triggered a glibc bug on x86-64 that was
fixed only on glibc 2.33. So in order to use 2.31 we need to backport
the patch.

Reference: https://sourceware.org/bugzilla/show_bug.cgi?id=27033