Fix pseudo crash when calling basename function
This could make pseudo to seg fault in some conditions. The problem is
two conflicting patches.

One of them is the one that adds support for a custom
PSEUDO_CHROOT_XTRANSLATION variable, which uses the basename function.

The basename function is either defined in <strings.h> when _GNU_SOURCE
is defined or <libgen.h> (POSIX version) when this header is included.

The patch alone would normally work, but this other patch from a more
recent poky version changes to _DEFAULT_SOURCE instead of _GNU_SOURCE:

https://git.yoctoproject.org/poky/commit/meta/recipes-devtools/pseudo/files/glibc238.patch?h=kirkstone&id=4682ae38f285f59b68196289ece5802460805201

So when using both patches the basename function wasn't even declared
(gcc gave a warning), causing a seg fault. Since when using poky
_GNU_SOURCE can't be used anymore, the only solution left is to include
the <libgen.h> header.