recipes-extended: xen: Remove cooling-maps from Linux DTS
The Linux thermal framework fails to register "a53_cores" thermal-zone
because it isn't able to register "cpu0" as a cooling device. Xen
modifies the CPU node, removing extra properties like "phandle" or
"#cooling-cells".  The thermal framework is looking for the cooling
device by phandle.

This commit removes "cooling-maps" from "a53_cores" thermal zone. In
this way, Linux will be able to register all thermal zones and measure
the temperature for each of them. Linux running over Xen will not be
able to use CPUs as cooling device, but will be able to shut down the
system in case the critical temperature is reached.

Issue: ALB-10663
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>