s32g274abluebox3: conf: Fix build of flashimage
Most S32CC targets pick all flash offsets now from the dtb.
The build fails if these offsets are not set there.
The mechanism is not currently used for a build on the BlueBox3 SOM,
so we provide the required offsets the traditional way, permitting
proper image builds like fsl-image-flash or fsl-image-base.

Issue: ALB-10761
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>