socketcan: canfdtest: Don't assume own frame will arrive before DUT frame
A frame is considered processed when it is marked as DONE. This implies
that the frame was sent (flag SENT is set), the ACK for the sent frame
was received (flag HOST_RECEIVED) and the DUT frame was received
(DUT_RECEIVED flag).
Before this commit, the code checked if a frame was processed (DONE)
only after a DUT frame is received. This doesn't take into account that
an ACK frame can be received after the DUT frame, case in which
'canfdtest' will remain blocked waiting for one more frame.

Issue: ALB-10677
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>