udev-extras: handling was broken for S32 platforms
Fixing LS side names via udev revealed a fundamental problem that needs
to be fixed first to get consistent udev behavior for the various image
types across platforms, also helping the s32g274abluebox3 target.

On S32 platforms MACHINE_EXTRA_RRECOMMENDS was incorrectly used. This
variable does not apply to core-images and was therefore never
functional. It measn that udev-extraconf was assumed to be applied but
it never happened. Given that udev-extraconf should be consistently
applied for Yocto images, it was moved into IMAGE_INSTALL within
fsl-image-base.

kernel-modules are already installed into the image, so that part does
not need duplication.

Issue: ALB-10792
Signed-off-by: Heinz Wrobel <Heinz.Wrobel@nxp.com>